refactor: split memory into focused modules
